Module G2.Scalar

type t = Fr.t

val order : Z.t

The order of the finite field

val size_in_bytes : int

minimal number of bytes required to encode a value of the field.

val check_bytes : Bytes.t -> bool

check_bytes bs returns true if bs is a correct byte representation of a field element

val zero : t

The neutral element for the addition

val one : t

The neutral element for the multiplication

val add : t -> t -> t

add a b returns a + b mod order

val mul : t -> t -> t

mul a b returns a * b mod order

val eq : t -> t -> bool

eq a b returns true if a = b mod order, else false

val negate : t -> t

negate x returns -x mod order. Equivalently, negate x returns the unique y such that x + y mod order = 0

val inverse_opt : t -> t option

inverse_opt x returns x^-1 if x is not 0 as an option, else None

val pow : t -> Z.t -> t

pow x n returns x^n

val of_bytes_opt : Bytes.t -> t option

From a predefined bytes representation, construct a value t. It is not required that to_bytes (Option.get (of_bytes_opt t)) = t. By default, little endian encoding is used and the given element is modulo the prime order

val to_bytes : t -> Bytes.t

Convert the value t to a bytes representation which can be used for hashing for instance. It is not required that to_bytes (Option.get (of_bytes_opt t)) = t. By default, little endian encoding is used, and length of the resulting bytes may vary depending on the order.