Module T.Equality_witness

type (_, _) eq =

| Refl : ('a, 'a) eq

A proof witness that two types are equal.

type 'a t

A dynamic representation for 'a.

val make : unit -> 'a t

make () is a dynamic representation for 'a. A fresh identifier is returned each time make () is evaluated.

val eq : 'a t -> 'b t -> ('a, 'b) eq option

eq ida idb returns a proof that a = b if ida and idb identify the same type.

val hash : 'a t -> int

hash id returns a hash for id.